+++ src/main/java/com/malabudi/dineupbe/menu/repository/MenuGroupRepository.java
@@ -7,4 +7,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface MenuGroupRepository extends JpaRepository<MenuGroup, Integer> { }
+public interface MenuGroupRepository extends JpaRepository<MenuGroup, Long> { }
