+++ src/main/java/com/malabudi/dineupbe/menu/repository/MenuItemRepository.java
@@ -3,4 +3,4 @@ package com.malabudi.dineupbe.menu.repository;
 import com.malabudi.dineupbe.menu.model.MenuItem;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MenuItemRepository extends JpaRepository<MenuItem, Integer> { }
+public interface MenuItemRepository extends JpaRepository<MenuItem, Long> { }
